Add typings for React 18 alpha
