+++ types/react-dom/index.d.ts
@@ -10,6 +10,10 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.8
 
+// NOTE: Users of the React 18 alpha should add a reference
+// to 'react-dom/alpha' in their project. See alpha.d.ts's top comment
+// for reference and documentation on how exactly to do it.
+
 // NOTE: Users of the `experimental` builds of React should add a reference
 // to 'react-dom/experimental' in their project. See experimental.d.ts's top comment
 // for reference and documentation on how exactly to do it.
