+++ types/react-dom/tsconfig.json
@@ -2,7 +2,8 @@
     "files": [
         "index.d.ts",
         "test/react-dom-tests.tsx",
-        "test/experimental-tests.tsx"
+        "test/experimental-tests.tsx",
+        "test/alpha-tests.tsx"
     ],
     "compilerOptions": {
         "module": "commonjs",
