+++ types/react-is/alpha.d.ts
@@ -1,0 +1,37 @@
+/**
+ * These are types for things that are present in the React 18 alpha.
+ *
+ * Once React 18 is released they can just be moved to the main index file.
+ *
+ * To load the types declared here in an actual project, there are three ways. The easiest one,
+ * if your `tsconfig.json` already has a `"types"` array in the `"compilerOptions"` section,
+ * is to add `"react-is/alpha"` to the `"types"` array.
+ *
+ * Alternatively, a specific import syntax can to be used from a typescript file.
+ * This module does not exist in reality, which is why the {} is important:
+ *
+ * ```ts
+ * import {} from 'react-is/alpha'
+ * ```
+ *
+ * It is also possible to include it through a triple-slash reference:
+ *
+ * ```ts
+ * /// <reference types="react-is/alpha" />
+ * ```
+ *
+ * Either the import or the reference only needs to appear once, anywhere in the project.
+ */
+
+// See https://github.com/facebook/react/blob/master/packages/react-is/src/ReactIs.js to see how the exports are declared,
+
+import ReactIs = require('.');
+import { ReactElement } from 'react';
+
+export {};
+
+declare module '.' {
+    function isSuspenseList(value: any): value is ReactElement;
+
+    const SuspenseList: symbol;
+}
