+++ types/react-is/tsconfig.json
@@ -20,6 +20,7 @@
     },
     "files": [
         "index.d.ts",
-        "react-is-tests.tsx"
+        "test/react-is-tests.tsx",
+        "test/alpha-tests.tsx"
     ]
 }
