+++ types/react/OTHER_FILES.txt
@@ -1,3 +1,4 @@
+alpha.d.ts
 experimental.d.ts
 jsx-dev-runtime.d.ts
 jsx-runtime.d.ts
