+++ types/react/index.d.ts
@@ -30,6 +30,10 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.8
 
+// NOTE: Users of the React 18 alpha should add a reference
+// to 'react/alpha' in their project. See alpha.d.ts's top comment
+// for reference and documentation on how exactly to do it.
+
 // NOTE: Users of the `experimental` builds of React should add a reference
 // to 'react/experimental' in their project. See experimental.d.ts's top comment
 // for reference and documentation on how exactly to do it.
