+++ types/react/tsconfig.json
@@ -7,7 +7,8 @@
         "test/elementAttributes.tsx",
         "test/managedAttributes.tsx",
         "test/hooks.tsx",
-        "test/experimental.tsx"
+        "test/experimental.tsx",
+        "test/alpha.tsx"
     ],
     "compilerOptions": {
         "module": "commonjs",
